+++ packages/graphql/lib/schema-builder/collections/field-directive.collection.ts
@@ -1,4 +1,4 @@
-import { MetadataListByNameCollection } from './metadata.list.by.name.collection';
+import { MetadataListByNameCollection } from './metadata-list-by-name.collection';
 import { PropertyDirectiveMetadata } from '../metadata';
 
 export class FieldDirectiveCollection extends MetadataListByNameCollection<PropertyDirectiveMetadata> {
@@ -6,13 +6,14 @@ export class FieldDirectiveCollection extends MetadataListByNameCollection<Prope
   fieldNames = new Set<string>();
 
   add(value: PropertyDirectiveMetadata) {
-    if (this.sdls.has(value.sdl) && this.fieldNames.has(value.fieldName))
+    if (this.sdls.has(value.sdl) && this.fieldNames.has(value.fieldName)) {
       return;
+    }
 
     super.add(value, value.fieldName);
 
     this.sdls.add(value.sdl);
     this.fieldNames.add(value.fieldName);
-    this.globalArray && this.globalArray.push(value);
+    this.globalArray?.push(value);
   }
 }
