+++ packages/graphql/lib/schema-builder/collections/index.ts
@@ -1,7 +1,7 @@
 export * from './metadata.storage.collection.list';
-export * from './field.directive.collection';
+export * from './field-directive.collection';
 export * from './array.collection';
 export * from './metada.collection.model.interface';
-export * from './metadata.by.name.collection';
-export * from './metadata.list.by.name.collection';
+export * from './metadata-by-name.collection';
+export * from './metadata-list-by-name.collection';
 export * from './metadata.storage.collection';
