+++ packages/graphql/lib/schema-builder/collections/metadata-by-name.collection.ts
